+++ code/include/MIP.hpp
@@ -9,6 +9,7 @@ using namespace Utils;
 #define CPLEX_LOG_DIR "../log/cplex_out/log/"
 #define MIP_LOG_DIR "../log/cplex_out/mip/"
 #define INST_DIR "../data/"
+#define CPLEX_CORE 1
 
 #define MIP_DUAL_PRIM_GAP_TOL 1e-4
 #define MIP_GAP_TOL 0.0
@@ -70,9 +71,11 @@ public:
 	std::vector<double> getSol();
 
 	[[nodiscard]]
-	inline int getNumCols() noexcept { return CPXgetnumcols(env, model); }; // num cols = num var
+	virtual int getMIPNumVars() noexcept { return CPXgetnumcols(env, model); }
 	[[nodiscard]]
-	inline int getNumRows() noexcept { return CPXgetnumrows(env, model); }; // num rows = num constr
+	inline int getNumCols() noexcept { return CPXgetnumcols(env, model); } // num cols = num var
+	[[nodiscard]]
+	inline int getNumRows() noexcept { return CPXgetnumrows(env, model); } // num rows = num constr
 
 	MIP& addCol(const std::vector<double>& newCol, const double objCoef, const double lb, const double ub, const std::string name);
 	MIP& addCol(const size_t index, const double value, const double objCoef, const double lb, const double ub, const std::string name);
@@ -95,7 +98,7 @@ public:
 	inline MIP& saveModel() noexcept {
 		CPXwriteprob(env, model, (MIP_LOG_DIR + fileName + "_" + id + ".lp").c_str(), NULL);
 		return *this;
-	};
+	}
 	inline MIP& saveLog() noexcept {
 		CPXsetlogfilename(env, (CPLEX_LOG_DIR + fileName + "_" + id + ".log").c_str(), "w");
 		return *this;