+++ code/include/OMIP.hpp
@@ -18,10 +18,9 @@ public:
 
 	[[nodiscard]]
 	double getSlackSum();
+	
 	[[nodiscard]]
-	std::vector<double> getSol();
-	[[nodiscard]]
-	int getMIPNumVars() { return MIPNumVars; };
+	int getMIPNumVars() noexcept { return MIPNumVars; };
 
 private:
 	void setup();